Rename families::command::workload::playlist
The families::command::workload::playlist module does not contain
anything pertaining to playlist, rename this module to `command_iter`.

Make command_iter private and re-export `CommandGeneratingIter`.
